Move scripts/release to cfme/utils/release
Update import in cfme/scripting/miq
