+++ cfme/utils/release.py
